refactor: add statusList interface to types folder
