+++ src/common/types/statusList.ts
@@ -1,0 +1,4 @@
+export interface StatusList {
+  bits: number;
+  lst: string;
+}
