+++ src/functions/issueHandler.ts
@@ -8,7 +8,8 @@ import { LogMessage } from "../logging/LogMessage";
 import { randomUUID } from "crypto";
 import { putObject } from "../common/aws/s3";
 import { getConfig } from "../config/getConfig";
-import { createToken, StatusList } from "../common/token/createToken";
+import { createToken } from "../common/token/createToken";
+import { StatusList } from "../common/types/statusList";
 
 const REQUIRED_ENV_VARS = [
   "SIGNING_KEY_ID",
